requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
six==1.17.0
sqlparse==0.5.3
